+++ hutool-core/src/test/java/cn/hutool/core/bean/BeanUtilTest.java
@@ -4,7 +4,6 @@ import cn.hutool.core.annotation.Alias;
 import cn.hutool.core.bean.copier.CopyOptions;
 import cn.hutool.core.bean.copier.ValueProvider;
 import cn.hutool.core.collection.CollUtil;
-import cn.hutool.core.lang.Console;
 import cn.hutool.core.map.MapUtil;
 import lombok.Getter;
 import lombok.Setter;
@@ -154,7 +153,6 @@ public class BeanUtilTest {
 		person.setSlow(true);
 
 		Map<String, Object> map = BeanUtil.beanToMap(person);
-		Console.log(map);
 		Assert.assertEquals("sub名字", map.get("aliasSubName"));
 	}